test(test-optimization): simplify Vitest EFD regression
